+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 3.0.0 (branch)
+* Fix: [#546] (https://github.com/savonrb/savon/issues/546) Making request with raw xml, shortcutting the build method: `operation.xml_envelope`
+
 * Fix: [#539] (https://github.com/savonrb/savon/issues/539) Implement element form 'qualified' or 'unqualified' correctly
 
 * This version requires Ruby 1.9.2 or higher. Ruby 1.8 is no longer supported!